Write basic outline, add CSS grid, start JS
Write skeleton outline with plan for JS execution. Also add
CSS grid to the etch-a-sketch-screen div. Learned how to work
with Grid a little better in the process. JS should set an HTML
style attribute for the grid. Finally, declared DOM constants for
DOM manipulation during the process.

At this juncture, it's actually able to populate the screen with
perfect squares. I'm really happy with this start.